backend/requirements.txt
web3==6.15.1
python-telegram-bot==20.7
anthropic==0.39.0
aiohttp==3.9.5
python-dotenv==1.0.1
aiosqlite==0.19.0
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
